pkg: Refactor model, modelmemory for game start feature

+++ pkg/model/room.go
@@ -14,10 +14,9 @@ type Room interface {
 
 	// Events returns feed of events happening in this room.
 	Events() RoomEventFeed
-
 }
 
 // RoomPayload represents payload of a game room.
 type RoomPayload struct {
-	MaxMemberCount uint
+	MaxMemberCount uint64
 }
